+++ tiny/src/components/image-gallery/toolbar/create-manifest/create-manifest-dialog.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
-import { useManifests } from '@/hooks/use-manifests';
+import { useManifests } from '@/hooks/use-manifest-metadata';
 import {
   Dialog,
   DialogContent,
