+++ tiny/src/components/image-gallery/toolbar/toolbar.tsx
@@ -8,7 +8,7 @@ import { DeleteSelectedButton } from './delete-selected';
 import { SelectAllButton } from './select-all';
 import { NavBreadcrumbs } from './nav-breadcrumbs';
 import { useUIState } from '@/hooks/use-ui-state';
-import { isManifest, isRootFolder } from '@/types';
+import { isManifestMetadata } from '@/types';
 import { ManifestActions } from '../manifest-actions';
 
 interface ToolbarProps {
@@ -37,7 +37,7 @@ export const Toolbar = (props: ToolbarProps) => {
 
         {!currentDirectory ? (
           <CreateManifestButton />
-        ) : isManifest(currentDirectory) ? (
+        ) : isManifestMetadata(currentDirectory) ? (
           <ManifestActions 
             manifest={currentDirectory} />
         ) : null}
