+++ tiny/src/layouts/base-layout.astro
@@ -1,7 +1,7 @@
 ---
-import '../styles/globals.css';
-
 import '@fontsource-variable/dm-sans';
+
+import '../styles/globals.css';
 
 const title = Astro.props.title || 'Minimal IIIF';
 ---
